refactor: 토픽 이름 변경
