+++ apps/commerce-streamer/src/main/java/com/loopers/interfaces/consumer/ProductEventConsumer.java
@@ -7,7 +7,7 @@ import org.springframework.kafka.support.Acknowledgment;
 import org.springframework.stereotype.Component;
 
 import com.loopers.confg.kafka.KafkaConfig;
-import com.loopers.domain.ProductPayload;
+import com.loopers.interfaces.consumer.dto.ProductPayload;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -15,16 +15,16 @@ import lombok.extern.slf4j.Slf4j;
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class ProductLikeConsumer {
+public class ProductEventConsumer {
 
     @KafkaListener(
-            topics = "productLike",
+            topics = "product-event",
             groupId = "loopers-default-consumer",
             containerFactory = KafkaConfig.BATCH_LISTENER
     )
     public void productLikeListener(List<ProductPayload> messages,
                                     Acknowledgment acknowledgment) {
-        messages.forEach(message -> log.info("productLike: {}", message));
+        messages.forEach(message -> log.info("product-event: {}", message));
         acknowledgment.acknowledge();
     }
 }
